start rsftecher test

+++ counterparty-core/counterpartycore/test/integrations/rsfetcher_test.py
@@ -1,66 +1,63 @@
+import os
 import shutil
-from time import sleep
+import tempfile
 
-import pytest
-from counterparty_rs import indexer
+from counterpartycore.lib import config
 from counterpartycore.lib.backend import rsfetcher
 
-TEST_DB_PATH = "/tmp/counterparty_test_db"  # noqa: S108
+TEST_DB_PATH = os.path.join(tempfile.gettempdir(), "rsfetcher_test_db")
+
+if os.path.exists(TEST_DB_PATH):
+    shutil.rmtree(TEST_DB_PATH)
 
 TEST_CONFIG = {
+    # "rpc_address": f"http://127.0.0.1:{PROXY_PORT}",
     "rpc_address": "http://127.0.0.1:8332",
     "rpc_user": "rpc",
     "rpc_password": "rpc",
     "db_dir": TEST_DB_PATH,
-    "log_file": "/Users/wilfred/Desktop/indexer_test.log",
+    "log_file": os.path.join(tempfile.gettempdir(), "rsfetcher_test.log"),
     "log_level": "debug",
     "json_format": False,
-    "consume_blocks": True,
+    "consume_blocks": False,
     "only_write_in_reorg_window": True,
-    "start_height": 868900,
 }
 
 
-@pytest.mark.skip()
 def test_fetcher_singleton():
-    fetcher = rsfetcher.RSFetcher(config=TEST_CONFIG)
-    fetcher.start()
-    fetcher.is_me = "Coucou"
-
-    fetcher2 = rsfetcher.RSFetcher(config=TEST_CONFIG)
-    fetcher2.start()
-    assert fetcher2.is_me == "Coucou"
-
-    assert fetcher.instance() is fetcher2.instance()
-    fetcher.stop()
-    shutil.rmtree(TEST_DB_PATH)
-
-
-@pytest.mark.skip()
-def test_fetcher_interrupt():
-    fetcher = rsfetcher.RSFetcher(config=TEST_CONFIG)
-    fetcher.start()
-    interrupted = False
+    config.NETWORK_NAME = "mainnet"
+    config.REGTEST = False
+    config.TESTNET3 = False
+    config.TESTNET4 = False
     try:
-        for _ in range(500):
-            block = fetcher.get_block()
-            assert isinstance(block, dict)
-            assert "height" in block
-            for t in block["transactions"]:
-                for v in t["vin"]:
-                    assert isinstance(v["script_sig"], bytes)
-    except KeyboardInterrupt:
-        interrupted = True
-        pass
+        # start_http_proxy("https://api.counterparty.io:8332")
+        fetcher = rsfetcher.RSFetcher(indexer_config=TEST_CONFIG)
+        fetcher.start(start_height=884596)
+        block = fetcher.get_block()
+        assert block["height"] == 884596
+        found_tx = None
+        for tx in block["transactions"]:
+            if tx["tx_id"] == "34e4856f35ac24dadea24e85478153921b1b341cf5207b0f1318947133ad6c57":
+                found_tx = tx
+        assert found_tx is not None
+
+        assert found_tx["parsed_vouts"] == (
+            ["1LnYFKsS575LqYKTqywaSgvVBe3NUxm673"],
+            546,
+            -5761,
+            b"\x16\x00\x00N\xc5O0;\xc2\x00\x00\x00\x00\x00\x00\x00\x00\x00\x01\x00",
+            [
+                ("1LnYFKsS575LqYKTqywaSgvVBe3NUxm673", 546),
+                (None, None),
+                ("bc1qvhxvtd5tznhh76ptwjv5x70qkyvm6swqch7yse", 5215),
+            ],
+        )
+
+        print(found_tx["vin"][0])
+
     finally:
         fetcher.stop()
-        shutil.rmtree(TEST_DB_PATH)
-
-    assert interrupted
-
-
-@pytest.mark.skip()
-def test_indexer():
-    i = indexer.Indexer(TEST_CONFIG)
-    i.start()
-    sleep(1000000000)
+        pass
+        # stop_http_proxy()
+    config.NETWORK_NAME = "regtest"
+    config.REGTEST = True
